Visnu's finance Agentic AI assistant/requirements.txt
  phi>=0.6.0
  openai
  python-dotenv
  yfinance
duckduckgo-search
phidata
fastapi
streamlit
 
  
